% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{choice}
\alias{choice}
\title{Check general choices}
\usage{
choice(choice, possible = "")
}
\arguments{
\item{choice}{The choice you want to analyze}

\item{possible}{What are good inputs?
choice()}
}
\description{
Check general choices
}
