% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSitemap.R
\name{downloadSitemap}
\alias{downloadSitemap}
\title{Function to download the Sitemap.xml of a given Webpage}
\usage{
downloadSitemap(sitemap)
}
\arguments{
\item{sitemap}{The Sitemap you want to download. This can also be a Index-Sitemap
downloadSitemap()}
}
\description{
This function allows to Download the Sitemap.xml of a given Webpage
}
\examples{
downloadSitemap("http://ohren-reinigen.de/sitemap.xml")
}
