% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBingResults.R
\name{getBingResults}
\alias{getBingResults}
\title{Function to retrive data from Google Suggest for a specific keyword}
\usage{
getBingResults(keyword, results = 10)
}
\arguments{
\item{keyword}{The initial keyword you want the results for}

\item{results}{Number of results you want to get. 10 is the Limit.
getBingResults()}
}
\description{
The function allows you to get the search results from Bing. The Function extratcts just the organic Search results. the Output is a Dataframe containing, Title,Description, URL and Landing Page.
}
\examples{
getBingResults("R Project")
}
