% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSuggest.R
\name{googleSuggest}
\alias{googleSuggest}
\title{Function to retrive data from Google Suggest for a specific keyword}
\usage{
googleSuggest(keyword, language = "en", walkThrough = FALSE)
}
\description{
This function allows you to get more longtail Keywords for a specific given Keyword. The Tool uses the Google Autocomplete function to retrieve this keywords.
}
