% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htag_count.R
\name{htag_count}
\alias{htag_count}
\title{Function to retrive the number of specified H-Tag for a given URL}
\usage{
htag_count(url, hTag = "h1")
}
\arguments{
\item{url}{The URL you want to get the H-Tags for}

\item{hTag}{The H-Tags you want to get (exampale = h2-Tags). Default is H1
htag_count()}
}
\description{
This function allows you to get the number of H-Tags you specified in the Call for a given URL.
}
