% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLdescription_length.R
\name{HTMLdescription_length}
\alias{HTMLdescription_length}
\title{Function to retrive the number of characters a Meta Description contains for a given URL}
\usage{
HTMLdescription_length(url)
}
\arguments{
\item{url}{character. The url you want to check.}
}
\description{
This function allows you to get the number of characters in a Meta Description for a given URL.
}
