% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLrobots.R
\name{HTMLrobots}
\alias{HTMLrobots}
\title{Function to retrive the Meta Robots for a given URL}
\usage{
HTMLrobots(url)
}
\description{
This function allows you to get the Meta Robots for a given URL.
}
