% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLtitle_length.R
\name{HTMLtitle_length}
\alias{HTMLtitle_length}
\title{Function to retrive the number of characters a Title-Tag contains for a given URL}
\usage{
HTMLtitle_length(url)
}
\arguments{
\item{url}{character. The url you want to check.}
}
\description{
This function allows you to get the number of characters in a Title-Tag for a given URL.
}
