% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{internal_all_file}
\alias{internal_all_file}
\title{Check Screaming-Frog Input Files}
\usage{
internal_all_file(input)
}
\arguments{
\item{input}{The input you want to analyze}
}
\description{
Check Screaming-Frog Input Files
}
\keyword{internal.}
\keyword{internal_all_file()}
