% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isIndexable.R
\name{isIndexable}
\alias{isIndexable}
\title{Function to check if a URL is indexable}
\usage{
isIndexable(url, bot = "googlebot", errorType = TRUE)
}
\description{
This function checks if a given URL is indexable. Therfore it checks different rules:
Status-Code: Does the URl gives a correct Status-Code (200)
Canonical-Link: Is the canonical Link Element implementet, so the URL can be indexed?
Meta-Robots: Is the Meta-Robots blocking the ULR?
robots.txt: Is the robots.txt blocking the URL from crawling?
}
