% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastCached.R
\name{lastCached}
\alias{lastCached}
\title{Function to get the last Google Cache Date for a URL}
\usage{
lastCached(url)
}
\description{
This function allows you to get the latest Google Cache Date for a given URL.
}
