% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkCount.R
\name{linkCount}
\alias{linkCount}
\title{Function to retrive the number of outgoing links for a given Url.}
\usage{
linkCount(url, uniqueLinks = FALSE)
}
\description{
This function allows you to get the number of outgoing links for a given URL.You can specify if you want to count same Links as one or not.
}
