% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileFriendly.R
\name{mobileFriendly}
\alias{mobileFriendly}
\title{Function to retrieve Data from Googles Mobile Friendly Testing API}
\usage{
mobileFriendly(domain, APIToken = "AIzaSyC1gUZEsqk-nny6f3KnJyTKuy30O6fpGCw")
}
\description{
This function allows you to retrive Data from Googles Mobile Friendly Test API. The Function retrives the data for a given Url. The API provides you a TRUE for "is mobile friendly" and FALSE for "is not mobile friendly"
}
