% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mozLinkMetrics.R
\name{mozLinkMetrics}
\alias{mozLinkMetrics}
\title{Function to retrive Link Metrics from Moz-API}
\usage{
mozLinkMetrics(page, Access_ID, Secret_Key, Scope = "page_to_page",
  Limit = 1, Filter = "", Sort = "", SourceCols = "536870916",
  TargetCols = "536870916", LinkCols = "")
}
\description{
This function allows you to get the Moz Link Metrics. You need a API Key to retrieve this Data. You can generate one for free here: https://moz.com/help/guides/moz-api/mozscape/getting-started-with-mozscape/create-and-manage-your-account At the Moment you just get the Data provided by a free Account.
}
