% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagesInIndex.R
\name{pagesInIndex}
\alias{pagesInIndex}
\title{Function to retrive the number of indexed URLs for a given Domain or URL-String}
\usage{
pagesInIndex(url)
}
\arguments{
\item{url}{The url you want to get the number of results for
pagesInIndex()}
}
\description{
This function allows you to get the number of indexed URLs for a given Domain or URL-String. The Function performs a Google-Site request to gets the value.
}
\examples{
pagesInIndex("https://www.r-project.org/")
}
