% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageSpeed.R
\name{pageSpeed}
\alias{pageSpeed}
\title{Function to retrieve Data from Googles Pagespeed Testing API}
\usage{
pageSpeed(url, APIToken = "AIzaSyC1gUZEsqk-nny6f3KnJyTKuy30O6fpGCw")
}
\arguments{
\item{url}{The Url you want to perform the test for.}

\item{APIToken}{Your API Token for this Google Project. Please use your own token if you perform a lot of calls.
pageSpeed()}
}
\description{
This function allows you to retrive Data from Googles Pagespeed Testing API. The Function retrives the data (a numeric value between 0 and 100) for a given Url.
}
