% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseCode.R
\name{responseCode}
\alias{responseCode}
\title{Function to get the Status Code of a given URL}
\usage{
responseCode(url)
}
\description{
This function allows you to get the the status Code of a given URL
}
