% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoDiver.R
\name{seoDiver}
\alias{seoDiver}
\title{Function to retrive data from the SEO Diver API}
\usage{
seoDiver(url, type = "Suchreichweite")
}
\description{
This function allows you to get Data from the SEO Diver API. http://de.seodiver.com/api Note: This Data is only available for DE,CH,AT Search Results. You get "Suchreichweite", "Suchwahrnehmung" and "Statische Sichtbarkeit"
}
