% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{seor_packages}
\alias{seor_packages}
\title{The member packages of the seor suite}
\usage{
seor_packages(include_optional = FALSE)
}
\arguments{
\item{include_optional}{Whether to append planned members (currently
\code{robotstxtr}) that are only attached when already installed.}
}
\value{
A character vector of package names.
}
\description{
The member packages of the seor suite
}
\examples{
seor_packages()

}
