% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedcount.R
\name{sharedcount}
\alias{sharedcount}
\title{Function to get the Social Shares for a given URL}
\usage{
sharedcount(url, key)
}
\description{
This function allows to get the shares of a given Domain into R. Therefore the function uses the Sharedcount-API. To use the function you have to set up a free API-Account here: https://www.sharedcount.com/
}
