% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{sitemapxml_check}
\alias{sitemapxml_check}
\title{Check if the User provided a correct Sitemap.xml}
\usage{
sitemapxml_check(sitemap)
}
\arguments{
\item{sitemap}{The sitemap you want to analyze
sitemapxml_check()}
}
\description{
Check if the User provided a correct Sitemap.xml
}
