% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactionalSuggest.R
\name{transactionalSuggest}
\alias{transactionalSuggest}
\title{Function to retrive data from Amazon or eBay Suggest function for a given Keyword}
\usage{
transactionalSuggest(keyword, language = "en_US", page = "amazon")
}
\arguments{
\item{keyword}{The initial keyword you want to get suggestions for}

\item{language}{The language you want to get suggestions. Default is en.}

\item{page}{The Page you want to get the Suggestions from amazon or eBay. Default is amazon.
googleSuggest()}
}
\description{
This function allows you to get more longtail Keywords for a specific given Keyword. The Tool uses the Google Autocomplete function to retrieve this keywords.
}
\examples{
\dontrun{
transactionalSuggest(keyword = "t shirt", language = "en_US", page = "amazon")
}
}
