% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{url_with_http}
\alias{url_with_http}
\title{URL Input checks}
\usage{
url_with_http(url)
}
\arguments{
\item{url}{The URL you want to analyze}
}
\description{
URL Input checks
}
\keyword{internal.}
\keyword{url_with_http()}
