% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipediaTraffic.R
\name{wikipediaTraffic}
\alias{wikipediaTraffic}
\title{Function to get Pageview Data of a given Wikipedia Article}
\usage{
wikipediaTraffic(page, start, end, project = "en.wikipedia.org",
  platform = "all-access", agent = "user", granularity = "daily")
}
\arguments{
\item{page}{The Page you want to get the traffic for for example cat.}

\item{start}{The start date. Format = YYYY-MM-DD}

\item{end}{The end date. Format = YYYY-MM-DD}

\item{project}{The Wikipedia Project you want to retrieve the data from. Default is en.wikipedia.org. A list of possible projects can be found here <https://meta.wikimedia.org/w/api.php?action=sitematrix&formatversion=2>}

\item{platform}{One of: all-access (default), desktop, mobile-app, mobile-web}

\item{agent}{One of 'user' (human reader, standard), 'spider' (search engine crawler), 'bot' (WMF bots) or 'all-agents'(user, spider and bot)}

\item{granularity}{The time unit for the response data. As of today, the only supported granularity for this endpoint is daily and monthly.
wikipediaTraffic()}
}
\description{
This function allows to get the pageviews for a given Wikiepdia article.
}
\examples{
wikipediaTraffic("R (programming language)",'2018-01-01','2018-01-10')
}
