% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allowedByRobots.R
\name{allowedByRobots}
\alias{allowedByRobots}
\title{Function to check if URL is blocked by robots.txt}
\usage{
allowedByRobots(url, bot = "googlebot")
}
\description{
This function checks if a given URL is blocked by the robots.txt file of the website.
}
