% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{choice}
\alias{choice}
\title{Check general choices}
\usage{
choice(choice, possible = "")
}
\arguments{
\item{choice}{character.}
\item{possible}{character.}
}
\description{
Check general choices
}
