% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainAge.R
\name{domainAge}
\alias{domainAge}
\title{Function to retrieve the Age of a given Domain}
\usage{
domainAge(domain)
}
\arguments{
\item{domain}{The domain you want to get the age for.
domainAge()}
}
\description{
This function allows you to get the Do´main Age for a given Domain. (Attention:At the moment just possible with .com-Domains)
}
