% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSitemap.R
\name{downloadSitemap}
\alias{downloadSitemap}
\title{Function to download the Sitemap.xml of a given Webpage}
\usage{
downloadSitemap(sitemap)
}
\arguments{
\item{sitemap}{character. The sitemap you want to download.}
}
\description{
This function allows to Download the Sitemap.xml of a given Webpage
}
