% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLinks.R
\name{extractLinks}
\alias{extractLinks}
\title{Function to extract all links of a given Url.}
\usage{
extractLinks(url, linkType = "all", uniqueLinks = FALSE)
}
\arguments{
\item{url}{character. The url you want to extract the Links from.}
\item{linkType}{character. What type of Links you want to check? ALL, or just "internal" or "external".}
\item{uniqueLinks}{logical. Extract unique Links as one is TRUE. Default is FALSE.}
}
\description{
This function allows you to extract all links of a given URL. You can specify if you just want unique Links or all. The output is a Data Frame.
}
