% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLinks.R
\name{extractLinks}
\alias{extractLinks}
\title{Function to extract all links of a given Url.}
\usage{
extractLinks(url, linkType = "all", uniqueLinks = FALSE)
}
\description{
This function allows you to extract all links of a given URL. You can specify if you just want unique Links or all. The output is a Data Frame.
}
