% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSuggest.R
\name{googleSuggest}
\alias{googleSuggest}
\title{Function to retrive data from Google Suggest for a specific keyword}
\usage{
googleSuggest(keyword, language = "en", walkThrough = FALSE)
}
\description{
This function allows you to get the Meta-Description of a specific URL.
}
