% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{host_tld}
\alias{host_tld}
\title{Check if the User just provided the Hostname and TLD}
\usage{
host_tld(url)
}
\arguments{
\item{url}{character. A url you want to check.}
}
\description{
Check if the User just provided the Hostname and TLD
}
