% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrefLang.R
\name{hrefLang}
\alias{hrefLang}
\title{Function to retrive the hrefLang-Tag for a given URL}
\usage{
hrefLang(url)
}
\arguments{
\item{url}{The URL you want to get the hrefLang-Tag for
hrefLang()}
}
\description{
This function allows you to get the hrefLang-Tags for a given URL. You get the Output as a Data Frame containing the Link and the Country/Language specififcation.
}
\examples{
hrefLang("https://www.r-project.org/")
}
