% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{htag_input}
\alias{htag_input}
\title{Check h-Tag Input}
\usage{
htag_input(h_tag)
}
\arguments{
\item{h_tag}{The h_tag you want to analyze}
}
\description{
Check h-Tag Input
}
\keyword{htag_input()}
\keyword{internal.}
