% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLcanonical.R
\name{HTMLcanonical}
\alias{HTMLcanonical}
\title{Function to retrive the Canonical Link element for a given URL}
\usage{
HTMLcanonical(url)
}
\arguments{
\item{url}{The URL you want to get the Canonical-Tag for
HTMLcanonical()}
}
\description{
This function allows you to get the canonical Link element for a given URL.
}
