% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLdescription.R
\name{HTMLdescription}
\alias{HTMLdescription}
\title{Function to retrive the Meta-Description of a given URL}
\usage{
HTMLdescription(url)
}
\arguments{
\item{url}{The URL you want to get the description for
HTMLdescription()}
}
\description{
This function allows you to get the Meta-Description of a specific URL.
}
\examples{
HTMLdescription("https://www.r-project.org/")
}
