% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLtitle.R
\name{HTMLtitle}
\alias{HTMLtitle}
\title{Function to retrive the Title of a given URL}
\usage{
HTMLtitle(url)
}
\arguments{
\item{url}{The URL you want to get the title for
HTMLtitle()}
}
\description{
This function allows you to get the title of a specific URL.
}
