% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{internal_all_file}
\alias{internal_all_file}
\title{Check Screaming-Frog Input Files}
\usage{
internal_all_file(input)
}
\arguments{
\item{input}{character. The input ScreamingFrog-File.}
}
\description{
Check Screaming-Frog Input Files
}
