% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordResults.R
\name{keywordResults}
\alias{keywordResults}
\title{Function to retrive the Number of results for a given Keyword}
\usage{
keywordResults(keyword, searchengine = "google")
}
\arguments{
\item{keyword}{The keyword you want to get the number of results for}

\item{searchengine}{The Searchengine you want to get the results form. Even Google or Bing. Default is Google.
keywordResults()}
}
\description{
This function allows you to get the number of Results for a given Keyword. You can get the value even from Google or Bing, by specifing this in the Call.
}
\examples{
keywordResults("R Project", searchengine = "google")
}
