% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mozUrlMetrics.R
\name{mozUrlMetrics}
\alias{mozUrlMetrics}
\title{Function to retrive URL Metrics from Moz-API}
\usage{
mozUrlMetrics(page, Access_ID, Secret_Key)
}
\arguments{
\item{page}{The page you want the data for}

\item{Access_ID}{Your Moz Access ID}

\item{Secret_Key}{Your Moz Secret Key
mozUrlMetrics()}
}
\description{
This function allows you to get the Moz URL Metrics. You need a API Key to retrieve this Data. You can generate one for free here: https://moz.com/help/guides/moz-api/mozscape/getting-started-with-mozscape/create-and-manage-your-account At the Moment you just get the Data provided by a free Account.
}
