% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagesInIndex.R
\name{pagesInIndex}
\alias{pagesInIndex}
\title{Function to retrive the number of indexed URLs for a given Domain or URL-String}
\usage{
pagesInIndex(url)
}
\description{
This function allows you to get the number of indexed URLs for a given Domain or URL-String. The Function performs a Google-Site request to get´s the value.
}
