% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageSpeed.R
\name{pageSpeed}
\alias{pageSpeed}
\title{Function to retrieve Data from Googles Pagespeed Testing API}
\usage{
pageSpeed(url, APIToken = "AIzaSyC1gUZEsqk-nny6f3KnJyTKuy30O6fpGCw")
}
\description{
This function allows you to retrive Data from Googles Pagespeed Testing API. The Function retrives the data (a numeric value between 0 and 100) for a given Url.
}
