% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screamingfrog_crawlVsSitemap.R
\name{screamingfrog_crawlVsSitemap}
\alias{screamingfrog_crawlVsSitemap}
\title{Function to compare the crawled URLs with the URLs in the sitemap.xml}
\usage{
screamingfrog_crawlVsSitemap(crawl, sitemap, deltaIn = "sitemap",
  checkImages = FALSE)
}
\description{
This function allows to compare the crawled URLs with the URLs in the sitemap.xml. The Output are the missing URLs - in the Sitemap or in the Crawl
}
