% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screamingfrog_internalPagerank.R
\name{screamingfrog_internalPagerank}
\alias{screamingfrog_internalPagerank}
\title{Function to calculate the internal PageRank of a crawled Page}
\usage{
screamingfrog_internalPagerank(crawl, domain = NULL)
}
\description{
This function allows you to calculate the internal PageRank of a given Domain. The calculation is based on a ScreamingFrog Crawl and the export "all_outlinks.csv".
}
