% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seoDiver.R
\name{seoDiver}
\alias{seoDiver}
\title{Function to retrive data from the SEO Diver API}
\usage{
seoDiver(url, type = "Suchreichweite")
}
\arguments{
\item{url}{character. The URL you want to get the Data for.}
\item{type}{character. The type of Data you want to get. Possible Inputs are: "Suchreichweite", "Suchwahrnehmung" and "Statische Sichtbarkeit".}
}
\description{
This function allows you to get Data from the SEO Diver API. http://de.seodiver.com/api Note: This Data is only available for DE,CH,AT Search Results. You get "Suchreichweite", "Suchwahrnehmung" and "Statische Sichtbarkeit"
}
