% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seor-package.R
\docType{package}
\name{seor-package}
\alias{seor}
\alias{seor-package}
\title{seor: install and load the SEO and developer tool suite}
\description{
\code{seor} is a metapackage. It does not define analysis functions of its own;
instead it bundles a curated set of focused packages so that a single
\code{install.packages("seor")} pulls them in and a single \code{library(seor)}
attaches them all, the way \code{tidyverse} does for its members.
}
\details{
Use \code{\link[=seor_packages]{seor_packages()}} to see the member packages and \code{\link[=seor_conflicts]{seor_conflicts()}} to
inspect function-name collisions on the search path.
}
\author{
\strong{Maintainer}: Bart Turczynski \email{b.turczynski@tidio.net}

Authors:
\itemize{
  \item Bart Turczynski \email{b.turczynski@tidio.net}
}

}
\keyword{internal}
