% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{url_with_http}
\alias{url_with_http}
\title{URL Input checks}
\usage{
url_with_http(url)
}
\arguments{
\item{url}{character. The url-Input you want to check.}
}
\description{
URL Input checks
}
