% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlInSitemap.R
\name{urlInSitemap}
\alias{urlInSitemap}
\title{Function to check if a given URL is in a XML-Sitemap}
\usage{
urlInSitemap(url, sitemap)
}
\description{
This function allows you to check if a given URL is found in a XML-Sitemap. For this you can also put in your Index-Sitemap.
}
