% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3cValidate.R
\name{w3cValidate}
\alias{w3cValidate}
\title{Function to preform a W3C Validation Test for a specific URL}
\usage{
w3cValidate(url)
}
\arguments{
\item{url}{character. The URL you want to check}
}
\description{
This function allows you to perform a W3C Validation test for a given URL. You will get the number of Errors and Warnings for the URL as Data Frame.
}
